Add better handling in scenario where Homebrew can't uninstall package.
Yesterday I installed the 'yarn' package via Homebrew. It has a
dependency on the 'node' package. I ran `bin/dev` today and it failed
without an error message. This is because the laptop script attempts to
uninstall Homebrew's 'node' package, but it could not because the 'yarn'
package has a dependency on it.

Prior to this commit, if Homebrew encounters an error, we redirect them
to `/dev/null`, which is why I didn't have any idea what the issue was.

With the changes in this commit, I added more logging when we uninstall
a Homebrew package and only redirect stdout (not stderr) to `/dev/null`.