Automatically create /usr/bin/n after installing 'n' package.
When running `/bin/dev` for the first time we install the 'n' package
which manages node.js versions. When 'n' is first run, it expects the
`/usr/local/n` directory to be created. This is currently not the case
so the user hits an error and has to manually create it themselves.

This updates our laptop script to automatically create this directory if
it's not already created.

Relevant Slack convo from tonight:

* https://kickstarter.slack.com/archives/C0PNB94BU/p1515537550000149